+++ src/false/false.rs
@@ -9,6 +9,6 @@
  * file that was distributed with this source code.
  */
 
-pub fn uumain(_: Vec<String>) -> int  {
+pub fn uumain(_: Vec<String>) -> isize {
     1
 }
